requirements.txt
python-dotenv==1.0.1
discord.py==2.4.0
requests==2.32.3
pandas==2.2.3